Rename s:flush_xp()
The previous name was s:send_xp(), but the function did not do the same
thing as codestats.send_xp() on the Python side. The s:flush_xp()
function clears the current pending buffer XP and adds it to the dict
waiting to be sent, but the send will happen later (either as scheduled,
or on exit).